env_process: Refactor kernel version verification step
KVM (kernel) version is checked and compared to the required version,
if there's one. That was done in the preprocess and postprocess
functions in virttest.env_process. Write a Setuper subclass that
implements that in the setup method and register the setuper in the
env_process setup_manager.

This is a patch from a larger patch series refactoring the env_process
preprocess and postprocess functions. In each of these patches, a
pre/post process step is identified and replaced with a Setuper subclass
so the following can finally be met:
    - Only cleanup steps of successful setup steps are run to avoid
      possible environment corruption or hard to read errors.
    - Running setup/cleanup steps symmetrically during env pre/post
      process.
    - Reduce explicit pre/post process function code length.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>