env_process: Refactor version info logging step
After checking all test-relevant package versions, the dict carrying
the information must be logged into the test data. That was done in the
preprocess and postprocess functions in virttest.env_process. Write a
Setuper subclass that implements that in the setup method and register
the setuper in the env_process setup_manager.

While on it, remove the requirement_checks import from env_process as
all steps are now refactored into Setupers in requirement_checks.

This is a patch from a larger patch series refactoring the env_process
preprocess and postprocess functions. In each of these patches, a
pre/post process step is identified and replaced with a Setuper subclass
so the following can finally be met:
    - Only cleanup steps of successful setup steps are run to avoid
      possible environment corruption or hard to read errors.
    - Running setup/cleanup steps symmetrically during env pre/post
      process.
    - Reduce explicit pre/post process function code length.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>