Use raw image format by default for block devices
Not providing a default behavior leads to use of variable without
a value and thus cryptic errors regarding the simple fact that we
have to provide an image format. It is also not recommended in any
conditionals as it leads to undefined behaviors. Finally, in the
case of multiple images we may have "image_format=" to use empty
string which should be handled similarly to None (lack of any
image_format parameter) or generally as the comment suggests to
a meaningful default behavior.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>