env_process: Don't raise TestFail in postprocess if hugepages leak
Postprocess checked whether hugepages were leaked after a related test
or not and raised a TestFail right after that.

During the refactoring of the logic in charge of setting up and cleaning
up huge pages for the test environment, a discussion was raised on
whether this should be tolerated or not (see [0]). The discussion
concluded that this should be avoided, even if the behavior of related
test cases in the event of huge pages leaks would change.

This commit removes the check from postprocess and adds it to the right
Setuper intended to run the huge page setup/cleanup steps. As the
TestFail would be masked by the RuntimeError that the SetupManager
raises in the end, this commit also raises a HugePagesError instead of
the former TestFail in favor of clarity, although it will be masked by
SetupManager's RuntimeError anyway.

[0] https://github.com/avocado-framework/avocado-vt/pull/4054

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>