env_process: Refactor kill_tail_threads
aexpect.kill_tail_threads was called right after the IPSniffer was
cleaned up. I missed refactoring kill_tail_threads when it was the right
time to do so. This patch fixes it by registering it in the same order
as before we started refactoring env_process.

This is a patch from a larger patch series refactoring the env_process
preprocess and postprocess functions. In each of these patches, a
pre/post process step is identified and replaced with a Setuper subclass
so the following can finally be met:
    - Only cleanup steps of successful setup steps are run to avoid
      possible environment corruption or hard to read errors.
    - Running setup/cleanup steps symmetrically during env pre/post
      process.
    - Reduce explicit pre/post process function code length.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>