pr-manager: Suppor persistent reservation manager
Persistent reservation managers:
SCSI persistent reservations allow restricting access to block devices to
specific initiators in a shared storage setup.

Defining a persistent reservation manager:
A persistent reservation manager is an instance of a subclass of the
“pr-manager” QOM class.
The pr-manager-helper which forwards the commands to an external privileged
helper program over Unix sockets. It has a single string property, path,
which accepts the path to the helper program’s Unix socket.
For example, the following command line defines
a pr-manager-helper object and attaches it to a SCSI passthrough device:

$ qemu-system-x86_64
-device virtio-scsi
-object pr-manager-helper,id=helper0,path=/var/run/qemu-pr-helper.sock
-drive if=none,id=hd,driver=raw,file.filename=/dev/sdb,file.pr-manager=helper0
-device scsi-block,drive=hd

Alternatively, using -blockdev:
$ qemu-system-x86_64
-device virtio-scsi
-object pr-manager-helper,id=helper0,path=/var/run/qemu-pr-helper.sock
-blockdev node-name=hd,driver=raw,file.driver=host_device,\
file.filename=/dev/sdb, file.pr-manager=helper0
-device scsi-block,drive=hd

You will also need to ensure that the helper program qemu-pr-helper is
running, and that it has been set up to use the same socket filename as
your QEMU commandline specifies.

Refercence:
https://www.qemu.org/docs/master/system/pr-manager.html
https://www.qemu.org/docs/master/tools/qemu-pr-helper.htm

Signed-off-by: Yongxue Hong <yhong@redhat.com>