utils_misc:add ID_SCSI_SERIAL into get_linux_drive_path search field
The serial and device can be set both on scsi-hd.

-device '{"driver": "scsi-hd", ..., "serial": "stg5_serial", "device_id": "stg5_device"}' \

The device_id is mapped to ID_SERIAL
The serial is mapped to ID_SCSI_SERIAL in the udevadm info output

Currently we miss ID_SCSI_SERIAL in the search field.
It will fail if we want to find a disk by serial instead of device_id

Signed-off-by: qingwangrh <qinwang@redhat.com>