iscsi: Adapt package names for SUSE installations
When attempting to install iSCSI initiator and target packages on SUSE,
it was observed that the test suite was failing due to incorrect
package names.

Specifically, SUSE utilizes `open-iscsi` for the iSCSI initiator,
as opposed to `iscsi-initiator-utils`.
Similarly, the SCSI target utilities on SUSE are provided by the
`tgt` package, not `scsi-target-utils`.

This commit introduces a conditional check for SUSE distributions
to ensure the correct iSCSI initiator (`open-iscsi`) and target (`tgt`)
packages are selected and installed, thereby resolving installation
failures on SUSE systems.

Signed-off-by: Kowshik Jois B S <kowsjois@linux.ibm.com>