bootstrap: Add multi-host setup via cluster config
This change introduces support for setting up a multi-host testing
environment directly through the vt-bootstrap command.

A new command-line option, --vt-cluster-config, has been added. This
option accepts a path to a JSON file that defines the cluster topology,
including the hosts and a central controller.

The bootstrap process now includes steps to:
 - Parse the provided cluster configuration file.
 - Register each host as a node in the cluster, preparing its agent
   environment.

Signed-off-by: Yongxue Hong <yhong@redhat.com>
Co-authored-by: Xu Han <xuhan@redhat.com>
Co-authored-by: Zhenchao Liu <zhencliu@redhat.com>