Reduction of Cirrus CI tasks
This will reduce the number of CI task which we run in Cirrus CI. It is
needed for avoiding the Cirrus CI's computation limit in the future. It
removes the tests with avocado 92lts version, which is not used anymore.
Also, it removes the `setup.py` installation method which is planned to
be removed soon, because the `setup.py` has been deprecated in python.

Signed-off-by: Jan Richter <jarichte@redhat.com>