docs: add DeepWiki reference as additional resource
The DeepWiki link is added as a new troubleshooting option in the Get
Started guide and as a technical reference in the Contributing
Guidelines. This provides users with more comprehensive documentation
and self-help options before reaching out to the community through
mailing lists, issues, or IRC. The DeepWiki contains in-depth technical
information about the Avocado framework.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>