netperf_stress: Support running netperf clients on systems with more than 256 CPUs (#4062)
* utils_netperf: split netperf setup and compilation

When compiled, netperf source code is decompressed first, then it is
compiled. However, both of these steps happen under the same subprocess
command.

There could be some cases in which it would be valuable to sepparate
those steps. For example, if some further netperf source code
modifications were needed.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>

* netperf: Support modifying max CPUs

netperf supports systems with up to 256 CPUs by default. There are,
though, larger systems with more CPUs that netperf struggles to work on.
According to netperf's error traces, modifying the number of MAX CPUs
previous to compilation is encouraged.

This patch adds the possibility of modifying the source code if needed.
To do so, it locates which is the file defining the macro (although the
default is src/netlib.h) and the value assigend to it. Then, it checks
the number of cpus the system running netperf has. If the number of cpus
exceeds the configured value of MAXCPUS, then it modifies the value
prior to compilation.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>

---------

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>