env_process: allow creating the dir while it exists
sometimes there is already the dir created and not mounted or
anything else before the test execution. And currently though
the code checks os.path.exists(), but while removing the dir,
it ignore_errors=True, which means that whatever the remove
sucesses or not, it will keep on running. So os.makedir() fails
for the dir exists sometimes.
so adding exist_ok=True for os.makedir() will allow dir creation
althrough the dir exits.

Signed-off-by: Tingting Mao <timao@redhat.com>