Getting ipv4 address from domiffadress
Few of the times it is seen that the ip address is not being fetch and the avocado runs were failed with error "ERROR: Failures occurred while postprocess:\n\n: Guest virt-tests-vm1 dmesg verification failed: Login timeout expired (output: 'exceeded 240 s timeout, last failure: No ipv4 DHCP lease for MAC aa:bb:cc:dd:ee:ff') "

To handle this error the patch has been sent. The patch helps in obtaining ip address of the guest using "virsh-net-dhcp-leases default" command. If the guest mac address is found in the command output, the mac ipv4 address is obatined and updated in the address.cache

If the ip_version is ipv6, then it raises exception VMIPAddressMissingError, and going to retry logic to try getting the ip address.
Even for ipv4 , if address is not found then it raises exception VMIPAddressMissingError and then retries will happen.

Signed-off-by: Tasmiya Nalatwad <tasmiya@linux.vnet.ibm.com>