Allow for configurable extra vm dump checks for present/absent messages
The typical use for this is when dumping extra logs and wanting to
make sure certain information is always present (e.g. startup, etc.)
and certain information is always absent (e.g. segfaults, etc.).

It makes use of aexpect's linux ops as well as pattern matrching
for grepping the expected or unexpected messages.

Signed-off-by: Plamen Dimitrov <plamen.dimitrov@intra2net.com>