virttest/staging/lv_utils: ignore warnings
The current implementation filters the data through grep. However, in
some cases a warning message might be displayed, such as

"""
File descriptor 3 (/dev/z90crypt) leaked on pvs invocation. Parent PID 5118: -bash
"""

While this is a valuable piece of information, the test case shouldn't break on
it if the command succeeds and the necessary information is found in the
output; the function focuses on identifying the blk target.

So, instead, read the pvs output and identify error conditions (command
failure, empty output). If these are not given, identify the relevant
parts as before from those output lines, that is the code does the
filtering, not grep.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>