utils_libvirt/libvirt_service: wait for service to terminate on kill
The function only issued the kill command but didn't wait for it to
become effective. This lead to timing issues with test steps afterwards,
possibly requiring invocation of `sleep` with absolute times.

Instead, now wait for a set timeout for the process to terminate. If it
doesn't, log this for better debugging later.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>